Add app settings using env variables
discussions/settings.py contains the production configuration,
and reads the settings of the app from the environment.
In our case the variables are set directly from CapRover.

discussions/settings_dev.py contains settings necessary for
development and it's enabled automaticcaly if the environment
variable DJANGO_DEVELOPMENT is present.

A Makefile is added to automate common actions.